updated awhole for msft_api.py
